requirements.txt
# Basic requirements for all
pre-commit 
black
flake8
regex
boto3

# Web scraping
requests
beautifulsoup4
lxml

# Data science
scikit-learn

# DB and data
pandas

# Other
jupyter
python-dotenv
